+++ zio-http/src/main/scala/zio/http/api/APIAddress.scala
@@ -1,0 +1,8 @@
+package zio.http.api
+
+sealed trait APIAddress
+object APIAddress {
+  def apply(server: String, port: Int): APIAddress = APIAddress.ServerPort(server, port)
+
+  final case class ServerPort(server: String, port: Int) extends APIAddress
+}
