+++ zio-http/src/main/scala/zio/http/api/APILocator.scala
@@ -1,0 +1,5 @@
+package zio.http.api
+
+trait APILocator {
+  def locate(api: API[_, _]): Option[APIAddress]
+}
