+++ zio-http/src/main/scala/zio/http/api/HeaderInputs.scala
@@ -1,3 +1,6 @@
 package zio.http.api
 
-private[api] trait HeaderInputs {}
+private[api] trait HeaderInputs {
+  def header[A](name: String, value: TextCodec[A]): In[A] =
+    In.Header(name, value)
+}
