--- zio-http/src/main/scala/zio/http/api/ServiceAddress.scala
@@ -1,8 +1,0 @@
-package zio.http.api
-
-sealed trait ServiceAddress
-object ServiceAddress {
-  def apply(server: String, port: Int): ServiceAddress = ServiceAddress.ServerPort(server, port)
-
-  final case class ServerPort(server: String, port: Int) extends ServiceAddress
-}
